fix some bug in tpcc grain

+++ Utilities/WorkloadMessageContents.cs
@@ -12,7 +12,7 @@ namespace Utilities
         public const int BASE_NUM_MULTITRANSFER = 150000;
         public const int BASE_NUM_NEWORDER = 150000;
 
-        public const double NUM_W_PER_4CORE = 4;
+        public const int NUM_W_PER_4CORE = 4;
         public const int NUM_D_PER_W = 10;
         public const int NUM_C_PER_D = 3000;
         public const int NUM_I = 100000;
